refactor: deep-review cleanups — shared test seeders, single-query cover check
Two findings from the final review pass: the attach and merge test
suites each carried their own seed_ebook/seed_audiobook/count helpers
(rule 03 bans duplicated seeders across files — they now live in
test_support as seed_synced_ebook/seed_synced_audiobook/count_rows),
and merge's adopt_cover_flag made two round-trips for two scalars that
one SELECT with two subqueries returns.
